Split the private_chef_ha LWRP back out into recipes plus a helper module
