+++ cookbooks/ec-harness/recipes/ec2.rb
@@ -1,7 +1,6 @@
 # encoding: utf-8
 
-require 'cheffish'
-require 'chef_metal_fog'
+require 'chef/provisioning/fog_driver/driver'
 
 harness_dir = node['harness']['harness_dir']
 repo_path = node['harness']['repo_path']
@@ -13,13 +12,9 @@ with_chef_local_server :chef_repo_path => repo_path,
     :port => 9010.upto(9999)
 
 with_driver "fog:AWS:default:#{node['harness']['ec2']['region']}"
-# alternative method:
-# with_driver 'fog:AWS:default', :compute_options => {
-#   :region => node['harness']['ec2']['region'],
-# }
 
-with_machine_options :ssh_username => node['harness']['ec2']['ssh_username'],
-  :use_private_ip_for_ssh => node['harness']['ec2']['use_private_ip_for_ssh']
+# with_machine_options :ssh_username => node['harness']['ec2']['ssh_username'],
+#   :use_private_ip_for_ssh => node['harness']['ec2']['use_private_ip_for_ssh']
 
 
 # override all keypair settings if passed as env var
@@ -32,39 +27,3 @@ fog_key_pair keypair_name do
   public_key_path File.join(repo_path, 'keys', 'id_rsa.pub')
 end
 
-# set provisioner options for all of our machines
-topo = TopoHelper.new(ec_config: node['harness']['vm_config'])
-topo.merged_topology.each do |vmname, config|
-  fog_helper = FogHelper.new(ami: node['harness']['ec2']['ami_id'], region: node['harness']['ec2']['region'])
-
-  local_provisioner_options = {
-    :ssh_username => config['ssh_username'] || node['harness']['ec2']['ssh_username'],
-    :bootstrap_options => {
-      :key_name => keypair_name,
-      :flavor_id => config['instance_type'] || 'm3.xlarge',
-      :region => node['harness']['ec2']['region'],
-      :ebs_optimized => config['ebs_optimized'] || false,
-      :image_id => config['ami_id'] || node['harness']['ec2']['ami_id'],
-      :subnet_id => config['vpc_subnet'] || node['harness']['ec2']['vpc_subnet'],
-      :associate_public_ip => true,
-      :block_device_mapping => [
-        {'DeviceName' => fog_helper.get_root_blockdevice,
-          'Ebs.VolumeSize' => 12,
-          'Ebs.VolumeType' => 'gp2',
-          'Ebs.DeleteOnTermination' => "true"},
-        {'DeviceName' => '/dev/sdb', 'VirtualName' => 'ephemeral0'}
-      ],
-      :tags => {
-        'EcMetal' => node['harness']['ec2']['ec_metal_tag']
-      }
-    }
-  }
-
-  node.set['harness']['provisioner_options'][vmname] = local_provisioner_options
-end
-
-# Precreate cloud machines, for dynamic discovery later on
-ec_harness_private_chef_ha "cloud_create_for_EC2" do
-  action :cloud_create
-  not_if { node['recipes'].include?('ec-harness::cleanup') }
-end
