+++ cookbooks/ec-harness/recipes/pedant.rb
@@ -2,6 +2,16 @@
 
 include_recipe "ec-harness::#{node['harness']['provider']}"
 
-ec_harness_private_chef_ha "run_pedant_on_#{node['harness']['provider']}" do
-  action :pedant
+ecm_topo_chef.merged_topology.each do |vmname, config|
+
+  # Skip on non-bootstrap backend
+  # FIXME: this assumes no failover has occured, better to make it conditional
+  next if ecm_topo_chef.is_backend?(vmname) && ecm_topo_chef.is_ha? &&
+    config['bootstrap'] != true
+
+  machine_execute "run_pedant_on#{vmname}" do
+    command '/opt/opscode/bin/private-chef-ctl test'
+    machine vmname
+  end
+
 end
