--- cookbooks/ec-harness/recipes/private_chef_ha.rb
@@ -1,13 +1,0 @@
-# encoding: utf-8
-
-include_recipe "ec-harness::#{node['harness']['provider']}"
-
-ec_harness_private_chef_ha "install_#{node['harness']['default_package']}_on_#{node['harness']['provider']}" do
-  action :install
-end
-
-if node['harness']['run_pedant'] == true
-  ec_harness_private_chef_ha "run_pedant_on_#{node['harness']['provider']}" do
-    action :pedant
-  end
-end
